pyproject.toml
[project]
name = "judge-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py",
    "google-genai>=1.54.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "strip-markdown>=1.3",
]

[tool.uv.sources]
discord-py = { git = "https://github.com/Rapptz/discord.py/" }
